fix: Ensure proper type casting for audio frequency data
- Updated the `getByteFrequencyData` method call in `SacredViz.tsx` to explicitly cast `currentAudio.freqs` as `Uint8Array` for improved type safety and to prevent potential runtime errors.

Files Changed:
- apps/integrity-pulse/src/components/SacredViz.tsx

[GI: 0.995] [Consensus: AUREA/ATLAS] [Hash: SHA256]

+++ apps/integrity-pulse/src/components/SacredViz.tsx
@@ -392,7 +392,7 @@ export default function SacredViz({ mockData = true }: SacredVizProps) {
       const t = clock.getElapsedTime();
       const currentAudio = audioContextRef.current;
       if (currentAudio && currentAudio.analyser) {
-        currentAudio.analyser.getByteFrequencyData(currentAudio.freqs);
+        currentAudio.analyser.getByteFrequencyData(currentAudio.freqs as Uint8Array);
       }
 
       // camera orbit
